hardware_scanner/requirements.txt
psutil>=5.9.0
qrcode[pil]>=7.4.2
certifi>=2024.0.0
flet>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
sounddevice>=0.4.6
# Windows only:
# wmi>=1.5.1; sys_platform == "win32"
# pywin32>=306; sys_platform == "win32"
